// apis/air-control/src/lingo/core/entity/Entity.d.ts
export interface IFrom {
    /**
     * GQE FULL OUTER JOIN OTHER
     */
    fullJoin<IF extends IFrom>(right: IF): IJoinFields<IF>;
    /**
     * GQE INNER JOIN OTHER
     */
    innerJoin<IF extends IFrom>(right: IF): IJoinFields<IF>;
    /**
     * GQE LEFT JOIN OTHER
     */
    leftJoin<IF extends IFrom>(right: IF): IJoinFields<IF>;
    /**
     * GQE RIGHT JOIN OTHER
     */
    rightJoin<IF extends IFrom>(right: IF): IJoinFields<IF>;
}
/**
 * Concrete Entity joined via Many-To-One or One-To-Many relation used in the FROM clause.
 */
export interface IEntityRelationFrom {
}
/**
 * A concrete Generated Query Entity.
 */
export interface IQEntity {
    fullJoin<IF extends IFrom>(right: IF): IJoinFields<IF>;
    innerJoin<IF extends IFrom>(right: IF): IJoinFields<IF>;
    leftJoin<IF extends IFrom>(right: IF): IJoinFields<IF>;
    rightJoin<IF extends IFrom>(right: IF): IJoinFields<IF>;
}
export interface IQEntityInternal extends IQEntity {
    __driver__: IQEntityDriver;
}
export interface IQEntityDriver {
    utils: IUtils;
    airportDb: IAirportDatabase;
    dbEntity: DbEntity;
    dbRelation: DbRelation;
    currentChildIndex: number;
    entityFieldMap: {
        [propertyName: string]: IQOperableFieldInternal<any, JSONBaseOperation, any, any>;
    };
    entityRelations: IQInternalRelation<any>[];
    fromClausePosition: number[];
    idColumns: IQOperableFieldInternal<any, JSONBaseOperation, any, any>[];
    allColumns: IQOperableFieldInternal<any, JSONBaseOperation, any, any>[];
    relations: IQInternalRelation<any>[];
    joinType: JoinType;
    joinWhereClause: JSONBaseOperation;
    parentJoinEntity: IQEntityInternal;
    getInstance(): IQEntityInternal;
    getRelationJson(columnAliases: IFieldColumnAliases<any>): JSONRelation;
    getRootJoinEntity(): IQEntityInternal;
    isRootEntity(): boolean;
    join<IF extends IFrom>(right: IF, joinType: JoinType): IJoinFields<IF>;
}